refactor: split tar helper to own module
